refactor(main): モジュールに分解
